Start building libffi on Windows.
Building off of prior efforts by @mingwandroid and @scopatz, and cribbing a lot from

http://hostagebrain.blogspot.com/2015/06/building-libffi-on-windows-by-visual.html

The autotools and libtool can handle the MSVC toolchain in some fashions, but
for libffi we need to compile a .S assembler file, which ends up being tricky
to work with. The easiest approach is to use the custom compiler wrapper
scripts provided with libffi (`msvcc.sh`).