Fix the Windows package: provide a static library, not a DLL.
Well, it turns out that the attempt to build a libffi DLL didn't work. The
libffi source code isn't annotated with __declspec(dllexport) markers, so
exporting the correct symbols would be a pain. (The recent libffi 3.3 release
candidate should fix this, fortunately.) And the `.lib` file installed by
Libtool in the package turned out to be exactly the same file as the DLL,
not an import library at all. I attempted to use `LIB /DEF` to generate
an import library but it errored on the DLL executable format (message
like this: https://stackoverflow.com/q/9657040/3760486), so it seemed
wisest just to give up.

The static library wasn't trouble-free. It was necessary to remove the "-GL"
flag from the environment, however. The "global link optimization" mode that
it enables seems to break the creation of static libraries via "ar cru" -- if
you run `nm` on the resulting file, you get "unknown file format" for the
object files compiled with that option active. It was also necessary to hack
the header to remove __declspec(dllimport) tags that would be adopted by users
-- otherwise they would look for symbols named like `__imp_ffi_foo`. Note that
this hack is essentially what's recommended in the header itself.

I also switched the build script to remove `.la` files for all platforms,
since we're making a point of ditching them.